server/requirements.txt
# FastAPI stack
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiohttp>=3.9.1
python-dateutil>=2.8.2

# Data handling - Upgraded for Python 3.13
numpy>=1.26.4
pydantic>=2.7.0
# pydantic-core will be installed as a dependency of pydantic

# OpenAI + LangChain - Upgraded for Python 3.13
# OpenAI + LangChain - Upgraded for Python 3.13
google-generativeai>=0.8.3
tenacity>=8.2.0
langchain>=0.1.20
langchain-core>=0.1.52

# OCR pipeline (Tesseract only)
opencv-python-headless>=4.8.1.78
Pillow>=11.0.0
pytesseract>=0.3.10

# PDF processing
pdf2image>=1.17.0
PyPDF2>=3.0.1

# Optional helper - Upgraded for Python 3.13
python-Levenshtein>=0.25.0

# Ensures clean builds
setuptools>=68.0.0